Updated stream_command tests with async_asgi_testclient client that seems to work with streaming async
