+++ experiments/stream_command.py
@@ -1,13 +1,5 @@
 
-from fastapi import FastAPI
-from fastapi.responses import StreamingResponse
-import itertools
 
-import asyncio
-
-import uvicorn
-
-stream_yes = FastAPI()
 
 blocksize = int(1024 * 1024 / 4)
 
@@ -44,50 +36,3 @@ async def stream_command_async(proc, logger=None, blksize=blocksize):
         # from the os, so we have to catch and ignore this error
         pass
 
-
-@stream_yes.get("/stream_yes")
-async def get_stream_yes():
-    """
-    Run the yes command and stream results. Yes runs forever so this should be an
-    infinite stream.
-    """
-    cmd = ["yes"]
-    proc = await asyncio.create_subprocess_exec(*cmd,
-                                                stdout=asyncio.subprocess.PIPE)
-    stream_gen = stream_command_async(proc)
-    return StreamingResponse(stream_gen)
-
-@stream_yes.get("/stream_yes_truncate")
-async def get_stream_yes_truncate():
-    """
-    Run the yes command and stream results. Yes runs forever so this should be an
-    infinite stream.
-    """
-    yes_cmd = "yes | head -n1000"
-    cmd = ["bash", "-c", yes_cmd]
-    proc = await asyncio.create_subprocess_exec(*cmd,
-                                                stdout=asyncio.subprocess.PIPE)
-    stream_gen = stream_command_async(proc)
-    return StreamingResponse(stream_gen)
-
-
-@stream_yes.get("/stream_yes_fake")
-async def get_stream_yes_fake():
-    """
-    Returns an infinite stream of "y"
-    """
-    y_gen = itertools.repeat("y\n")
-    return StreamingResponse(y_gen)
-
-
-@stream_yes.get("/stream_yes_fake_truncate")
-async def get_stream_yes_fake_truncate():
-    """
-    Returns an infinite stream of "y"
-    """
-    y_gen = itertools.repeat("y\n", 1000)
-    return StreamingResponse(y_gen)
-
-
-if __name__ == "__main__":
-    uvicorn.run(stream_yes, host="0.0.0.0", port=8000)
